Prevent some/more/all? cases of splitter drifting upward

+++ source/org/miradi/views/umbrella/PersistentNonPercentageHorizontalSplitPane.java
@@ -27,7 +27,7 @@ public class PersistentNonPercentageHorizontalSplitPane extends PersistentHorizo
 	}
 	
 	@Override
-	protected double computePercentFromLocation(int location)
+	protected int computePercentFromLocation(int location)
 	{
 		return location + 1;
 	}
